% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCLE_correlation_matrix}
\alias{CCLE_correlation_matrix}
\title{Gene correlations in CCLE cancer cell lines}
\format{
A \code{matrix} object with 308 rows and 24327 columns.
\itemize{
\item Rows correspond to CT genes
\item Columns correspond to all genes from CCLE database
}
}
\source{
See \code{scripts/make_CCLE_correlation_matrix.R} for details.
}
\description{
Correlation coefficients between Cancer-Testis genes and all genes
found on the CCLE database.
}
\details{
Correlation coefficients (Pearson) between CT genes and all other
genes are given in the matrix. These correlation coefficients were
calculated using log transformed expression values from \code{CCLE_data}
(all cell lines).
}
