% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTdata.R
\name{CTdata}
\alias{CTdata}
\title{All CTdata datasets}
\usage{
CTdata()
}
\value{
A \code{data.frame} describing the data available in
\code{CTdata}.
}
\description{
This is the companion Package for \code{CTexploreR} containing omics
data to select and characterise CT genes.

Data come from public databases and include expression and
methylation values of genes in normal and tumor samples as well as
in tumor cell lines, and expression in cells treated with a
demethylating agent is also available.

The \code{\link[=CTdata]{CTdata()}} function returns a \code{data.frame} with all the
annotated datasets provided in the package. For details on these
individual datasets, refer to their respective manual pages.

See the vignette and the respective manuals pages for more details
about the package and the data themselves.
}
\examples{

CTdata()
}
\author{
Laurent Gatto
}
