% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{embryo_sce_Petropoulos}
\alias{embryo_sce_Petropoulos}
\title{Embryo scRNAseq}
\format{
A \code{SingleCellExperiment} object with 26178 rows and 1481 columns
\itemize{
\item Rows correspond to genes (gene names as rownames)
\item Columns correspond to cells
}
}
\source{
RPKM and metadata files were downloaded from
https://www.ebi.ac.uk/biostudies/files/E-MTAB-3929/
The data were converted in a SingleCellExperiment
(see \code{scripts/make_embryo_sce_Petropoulos.R} for details).
}
\description{
Human embryo single cell RNAseq data in RPKM from
`Single-Cell RNA-Seq Reveals Lineage and X Chromosome Dynamics in
}
\details{
Description of the colData:
\itemize{
\item Column \code{individual} gives the sample the cell is coming from.
\item Column \code{stage} specifies the stage of the early embryo.
\item Column \code{sex} is the sex inference made using the expression of 11
Y-linked genes, made for each day individually.
\item Column \code{ambigous} indicates if the inference of the embryo's sex was
ambigous due to some cells expression of the Y-linked genes.
}
}
