% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{embryo_sce_Zhu}
\alias{embryo_sce_Zhu}
\title{Embryo scRNAseq}
\format{
A \code{SingleCellExperiment} object with 26255 rows and 50 columns
\itemize{
\item Rows correspond to genes (gene names as rownames)
\item Columns correspond to cells
}
}
\source{
50 FPKM files were downloaded from GEO
(accession: GSE81233). The data were converted in a SingleCellExperiment
(see \code{scripts/make_embryo_sce_Zhu.R} for details).
}
\description{
Human embryo single cell RNAseq data in FPKM from
`Single Cell DNA Methylome Sequencing of Human Preimplantation
}
\details{
Description of the colData:
\itemize{
\item Column \code{embryo} gives the embryo the cell is coming from.
\item Column \code{stage} specifies the stage of the early embryo.
\item Column \code{sex} is the sex inference made using the expression of RPS4Y1. If
mean expression of RPS4Y1 is higher than 50 FPKM, the sample is male.
}
}
