% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FGC_sce}
\alias{FGC_sce}
\title{Fetal gonad scRNAseq}
\format{
A \code{SingleCellExperiment} object with 22496 rows and 10850 columns
\itemize{
\item Rows correspond to genes (gene names as rownames)
\item Columns correspond to cells
}
}
\source{
ee58527e-e1e4-465d-8dc8-800ee40f14f2.rds file dowloaded from
https://cellxgene.cziscience.com/collections/661a402a-2a5a-4c71-9b05-b346c57bc451Data.
The data were converted in a SingleCellExperiment
(see \verb{scripts/072_make_FGC_sce.R} for details).
}
\description{
Human fetal gonad single cell RNAseq data from
\verb{Single-cell roadmap of human gonadal development}
(Garcia-Alonso, Nature 2022)
}
\details{
Description of the colData:
\itemize{
\item Column \code{type} gives the gender and the cell type.
\item Column \code{stage} specifies if the cell type is "pre-meiotic" or "meiotic".
\item Column \code{germcell} is set to TRUE when the cell type is a germ cell.
}
}
