% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hESC_data}
\alias{hESC_data}
\title{Genes expression in hESC}
\format{
A \code{SummarizedExperiment} object with 24488 rows and 4 columns
\itemize{
\item Rows correspond to genes (ensembl_gene_id as rownames)
\item Columns correspond to hESC types
\item Expression data from the assay are TPM values
}
}
\source{
RNAseq fastq files were downloaded from Encode databas (see
\code{scripts/make_hESC_data.R} for details).
}
\description{
Gene expression data in human embryonic stem cells
}
\details{
The colData contains
\itemize{
\item Column \code{genotype} gives the sexual genotype of the cells
}
}
