% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HPA_cell_type_specificity}
\alias{HPA_cell_type_specificity}
\title{Cell type specificities (from HPA)}
\format{
A \code{tibble} object with 24516 rows and 4 columns.
\itemize{
\item Rows correspond to genes (ensembl_gene_id)
\item Columns give genes cell type specificities
}
}
\source{
\code{proteinatlas.tsv} was downloaded from the Human Protein Atlas
(https://www.proteinatlas.org)
See \verb{scripts/081_make_HPA_cell_type_specificity.R} for details.
}
\description{
Cell type specificities based on scRNAseq data from the
Human Protein Atlas (https://www.proteinatlas.org)
}
\details{
\itemize{
\item Column \code{HPA_RNA_single_cell_type_specific_nTPM} gives the cell types in
which genes were detected.
\item Column \code{not_detected_in_somatic_HPA} specifies if the gene is detected or
not in a somatic cell type based on the value of the column
\code{HPA_RNA_single_cell_type_specific_nTPM}. Genes are flagged as
\code{TRUE} if the \code{HPA_RNA_single_cell_type_specific_nTPM} column
indicates a germ cell or a placental specificity. Genes are flagged as
\code{FALSE} if they are expressed in any other somatic cell
type. \code{NA} is set when the original table from HPA had no values for that
gene.
}
}
