% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methylation_in_hESC}
\alias{methylation_in_hESC}
\title{Methylation of CpGs within all genes promoters in hESC}
\format{
A \code{RangedSummarizedExperiment} object with 4280098 rows and 3 columns
\itemize{
\item Rows correspond to CpGs (located within all genes promoters (TSS +- 5000
nt))
\item Columns correspond to hESC
\item Methylation values from WGBS data
\item rowRanges correspond to CpG positions
}
}
\source{
WGBS methylation data was downloaded from Encode
(see \code{scripts/make_methylation_in_hESC.R} for details).
}
\description{
Methylation values of CpGs located within all genes
promoters in human embryonic stem cells.
}
