% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methylation_in_tissues}
\alias{methylation_in_tissues}
\title{Methylation of CpGs within all genes promoters}
\format{
A \code{RangedSummarizedExperiment} object with 4280327 rows and 14 columns
\itemize{
\item Rows correspond to CpGs (located within all genes promoters (TSS +- 5000
nt))
\item Columns correspond to normal tissues
\item Methylation values from WGBS data
\item rowRanges correspond to CpG positions
}
}
\source{
WGBS methylation data was downloaded from Encode and from GEO
databases (see \code{scripts/make_methylation_in_tissues.R} for
details).
}
\description{
Methylation values of CpGs located within all genes
promoters in a set of normal tissues.
}
