% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oocytes_sce}
\alias{oocytes_sce}
\title{Oocytes scRNAseq}
\format{
A \code{SingleCellExperiment} object with 26500 rows and 899 columns
\itemize{
\item Rows correspond to genes(gene names as rownames)
\item Columns correspond to cells
}
}
\source{
GSE154762_hO_scChaRM_count_matix.txt.gz\verb{was downloaded from GEO (accession: GSE154762). The data were converted in a SingleCellExperiment (see}scripts/make_oocytes_sce.R` for details).
}
\description{
Human oocytes single cell RNAseq data from
\verb{Decoding dynamic epigenetic landscapes in human oocytes using single-cell multi-omics sequencing} (Yan et al. Cell Stem Cell 2021)
}
\details{
Description of the colData:
\itemize{
\item Column \code{type} gives the cell type.
\item Column \code{stage} specifies if the cell type is "pre-meiotic" or "meiotic".
\item Column \code{germcell} is set to TRUE when the cell type is a germ cell.
}
}
