% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_methylation}
\alias{TCGA_methylation}
\title{Methylation of all genes promoters in TCGA samples}
\format{
A \code{RangedSummarizedExperiment} object with 79445 rows and 3423
columns
\itemize{
\item Rows correspond to Infinium 450k probes
\item Columns correspond to samples
\item Methylation data from the assay are Beta values
\item Clinical information are stored in colData
\item Probe information (hg38 coordinates) are stored in rowRanges
}
}
\source{
SKCM, LUAD, LUSC, COAD, ESCA, BRCA and HNSC methylation data were
downloaded with TCGAbiolinks and subsetted to select probes located
in CT genes promoter regions (see
\code{scripts/make_TCGA_methylation.R} for details).
}
\description{
Methylation values of probes located within all genes
promoters in samples from TCGA (tumor and peritumoral samples)
}
