% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CT_methylation_in_tissues}
\alias{CT_methylation_in_tissues}
\title{Methylation of CpGs located within CT promoters in normal tissues.}
\format{
A \code{RangedSummarizedExperiment} object with 53770 rows and 14 columns
\itemize{
\item Rows correspond to CpGs (located within CT genes promoters)
\item Columns correspond to normal tissues
\item Methylation values from WGBS data
\item rowRanges correspond to CpG positions
}
}
\source{
WGBS methylation data was downloaded from Encode and from GEO
databases (see \code{inst/scripts/make_CT_methylation_in_tissues.R} for
details).
}
\description{
Methylation values of CpGs located within Cancer-Testis (CT)
promoters in a set of normal tissues.
}
