% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeTags}
\alias{makeTags}
\title{A short function that returns the default CTdata tags and, if
provided, additional data-specific tags.}
\usage{
makeTags(x = character())
}
\arguments{
\item{x}{\code{character()} containing specific tags. Default is an
empty vector.}
}
\value{
A \code{character} containing the default tags and optional
data-specific tags.
}
\description{
A short function that returns the default CTdata tags and, if
provided, additional data-specific tags.
}
\examples{

CTdata:::makeTags() ## only default tags

CTdata:::makeTags("myTag") ## one additional tag

CTdata:::makeTags(c("myTag", "myOtherTag")) ## two additional tag
}
